requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
scikit-learn==1.6.0
numpy==1.26.4
faiss-cpu==1.9.0
requests==2.32.3